requirements.txt
streamlit
pyttsx3
pillow
